scripts/requirements.txt
# Зависимости для scripts/*.py.
# Установка: pip install -r scripts/requirements.txt
langfuse>=2.0,<4.0
python-frontmatter>=1.1
PyYAML>=6.0
